# Conflicts:
#	app-common/pom.xml
#	app-main/src/main/java/com/groupwork/charchar/controller/AttractionsController.java
#	app-main/src/main/java/com/groupwork/charchar/service/impl/AttractionsServiceImpl.java

+++ app-main/src/main/java/com/groupwork/charchar/entity/AttractionsEntity.java
@@ -58,7 +58,7 @@ public class AttractionsEntity implements Serializable {
     /**
      *
      */
-    private Integer attrRating;
+    private Double attrRating;
     /**
      *
      */
